Fix: Componentised not server component
